% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{as.character.icdb_icd10}
\alias{as.character.icdb_icd10}
\title{Convert a list of ICD10 codes to a character vector
Invalid ICD codes are returned as NA.}
\usage{
\method{as.character}{icdb_icd10}(x, ...)
}
\arguments{
\item{x}{The icd10 vector object to convert}

\item{...}{Unused addition parameters}
}
\description{
Convert a list of ICD10 codes to a character vector
Invalid ICD codes are returned as NA.
}
