% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{build_object_tree}
\alias{build_object_tree}
\title{Turn on or off caching}
\usage{
build_object_tree(con, prefix)
}
\arguments{
\item{con}{The database connection object}

\item{prefix}{The starting prefix defining the root of the tree (ID type)}

\item{choice}{TRUE will turn on the cache, FALSE will turn it off}
}
\value{
Nested list structure containing the object tree
}
\description{
Use this function to turn on or off caching. If caching is
enabled, then query results will be saved behind the scenes.
If caching is disabled, queries will always return results
directly from the database. Caching is disabled by default.
You must explicitly call this function if you want to use it.
You only need to call this function once, when you load the
ICDB library using library(icdb).
}
\details{
If caching is enabled, and you perform exactly the same query
again (for example, by running the same script), then the
cached results will be used, which will speed up the running
of the script. This can make development and debugging easier,
because the turnaround time for running commands and scripts
is reduced.

If caching is disabled, then results will always come from the
database. This can make code take longer to run, but will
always guarantee that results are up to date.
}
