% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{class}
\name{Cache-class}
\alias{Cache-class}
\title{Class for managing a query results cache}
\description{
Class for managing a query results cache
}
\section{Slots}{

\describe{
\item{\code{connection}}{Microsoft SQL Server.}

\item{\code{config}}{list. Database connection information as a named list}

\item{\code{dsn}}{}
}}

