% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{Cache}
\alias{Cache}
\title{Global cache object, not accessible outside cache
Create a new cache object}
\usage{
Cache(path = "cache/")
}
\value{
A new Cache object
}
\description{
Global cache object, not accessible outside cache
Create a new cache object
}
