% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{cache_object}
\alias{cache_object}
\title{Global cache object, ideally not accessible to users of the package
(not achieved that yet). This should probably be an R6 class. An R4
class will not work because it is not mutable. To fix later.}
\format{
An object of class \code{list} of length 2.
}
\usage{
cache_object
}
\description{
Global cache object, ideally not accessible to users of the package
(not achieved that yet). This should probably be an R6 class. An R4
class will not work because it is not mutable. To fix later.
}
\keyword{datasets}
