% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{$,Database-method}
\alias{$,Database-method}
\title{Access a table in a Database object}
\usage{
\S4method{$}{Database}(x, name)
}
\arguments{
\item{x}{The database}

\item{name}{The table to access}
}
\value{
A dplyr::tbl data source wrapper
}
\description{
Use this function to access a table in the Database object. The Database
object behaves like a list, but the elements of the list are evaluated only
when this function is called. When a call like db$table_name is made, this
function creates the corresponding dplyr::tbl object and returns it. Since
this is not a particularly long operation, the results do not need to be
cached. Use the resulting object for any processing you can do with
dplyr::tbl, e.g. pipe it to some SQL-type queries and collect().
}
