% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{$,Table-method}
\alias{$,Table-method}
\title{Overload $ for Table so that it returns the dplyr::tbl}
\usage{
\S4method{$}{Table}(x, name)
}
\arguments{
\item{x}{The Table object (element in a Databases object)}

\item{name}{The table name to get}
}
\value{
The dplyr::tbl for the selected table
}
\description{
Get the table with the specified name as a dplyr::tbl
}
