% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{codes_from}
\alias{codes_from}
\title{Remap a column of codes to hierarchy of strings}
\usage{
codes_from(tbl, codes_files, name)
}
\arguments{
\item{tbl}{The tbl to process}

\item{codes_files}{A list of the names of codes configuration files}

\item{name}{The logical column name (unquoted, like in dplyr filter etc.)}
}
\value{
The tbl after reducing
}
\description{
Read ICD codes (or any other codes) from a codes configuration file and
remap the column values with readable hierarchical strings.
}
\details{
Note that this function only currently works on logical columns with
only one source column.

Note that this function is a second layer strategy, meaning it must be
applied only on the resulting logical column (not the column names
which have a trailing _N).
}
