% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{collect}
\alias{collect}
\title{Collect and cache SQL query results}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{The dplyr SQL query to collect}

\item{...}{Other arguments for dplyr::collect()}
}
\value{
The query results
}
\description{
This function is a replacement for the dplyr::collect() function
that is used to submit an SQL query to a database. This function
is called in the same way, by piping to collect(), but it caches
the query behind the scenes, so that the next time the query is
fetched, the results are available quicker.
}
