% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{count_type}
\alias{count_type}
\title{Count the number of occurances of type}
\usage{
count_type(x, type)
}
\arguments{
\item{x}{The icd10 vector to count}

\item{type}{Character, One of "C", "E", "X", "T"}
}
\value{
The total number of matching types
}
\description{
Valid type
}
