% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Database-class}
\alias{Database-class}
\title{Database class wrapping an SQL server connection}
\description{
Database class wrapping an SQL server connection
}
\section{Slots}{

\describe{
\item{\code{connection}}{Microsoft SQL Server.}

\item{\code{config}}{list. Database connection information as a named list}

\item{\code{dsn}}{Domain source name (Windows only)}

\item{\code{.Data}}{From the contained list}
}}

