% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Databases-class}
\alias{Databases-class}
\title{Databases class wrapping an SQL server connection}
\description{
Databases class wrapping an SQL server connection
}
\section{Slots}{

\describe{
\item{\code{connection}}{Microsoft SQL Server.}

\item{\code{config}}{list. Databases connection information as a named list}

\item{\code{dsn}}{Domain source name (Windows only)}

\item{\code{.Data}}{From the contained list}
}}

