% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_s4.R
\docType{class}
\name{DatabaseS4-class}
\alias{DatabaseS4-class}
\title{Database class wrapping an SQL server connection}
\description{
Database class wrapping an SQL server connection
}
\section{Slots}{

\describe{
\item{\code{connection}}{Microsoft SQL Server.}

\item{\code{config}}{numeric.}
}}

