% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_s4.R
\name{DatabaseS4}
\alias{DatabaseS4}
\title{Construct a Database object}
\usage{
DatabaseS4(data_source_name = NULL, db_config = NULL)
}
\arguments{
\item{data_source_name}{The data source name (DSN) for the database.  If
this argument is provided, it will be preferred to the config file.}

\item{db_config}{The absolute path of a configuration file (csv format)
containing database connection information and credentials. The default
value is the credential file stored in the inst/ directory.}
}
\value{
A new (S4) Database object
}
\description{
This object manages a database connection. The connection is configured
using either a data source name, or a json file which stores
configuration information and credentials. After this function has run
without errors, you should have a new Database object containing a valid
connection.
}
\details{
A data source name (DNS) is a string that refers to a connection set
using the "ODBC data sources" app on Windows (8). Open this app; in the
"User DSN" tab, add a new data source using "Add". Select "SQL Server",
and then follow through the wizard. Choose a name for the data source,
which is arbitrary -- this is the data source name. Set the "Server"
field to the server name (the same as in SQL Management Studio). In one
of the pages, make sure you pick a default database to connect to.
Finally, at the end of the wizard, check the connection. If it fails,
debug it before moving on to using this function.

If you cannot make the DSN work, you can also supply a configuration file,
in JSON format. Three keys are required: "driver" must be set to the
database driver (potentially "SQL Server"); "server" must be set the to
server name; and "database" must be set to the name of the database to
connect to. An example configuration file is in extdata/db_config.json.

doneness 0
}
