% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{drop_detail}
\alias{drop_detail}
\title{Drop detail from a clinical codes column}
\usage{
drop_detail(codes, level = 1)
}
\arguments{
\item{codes}{A list (such as a dataframe column) to drop detail from}

\item{level}{At which level to truncate the codes}
}
\value{
The list after dropping detail past the specified level
}
\description{
Use this function to set the level of detail for diagnosis code
columns. Codes columns contain strings of the form foo.bar.some.
This function can be used to truncate this value at an arbitrary
level of detail, so that foo.bar.some because foo.bar (for level
2).
}
\details{
This function can be used as part of a mutate call to create a
new column with more coarse-grained information about diagnosis
codes; for example: mutate(new_col = drop_detail(diagnosis, 2)).

Dropping detail makes it easy to get rid of "noise" in the diagnosis
codes, and group together related codes into the hierarchy defined
by the codes mapping file.
}
