% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{drop_tags}
\alias{drop_tags}
\title{Remove tags from a clinical codes column}
\usage{
drop_tags(codes)
}
\arguments{
\item{codes}{A list of code strings to drop tags from}
}
\value{
The list of codes without the tags
}
\description{
Drop tags from a list of clinical code strings. A general code string
is of the form foo.bar\<tag\>.some, where a tag can appear anywhere.
After calling this function, all the \<tag\> elements are removed.
}
\details{
This function can be used as part of a mutate call, such as
mutate(new_col = drop_tags(codes_column))
}
