% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{gen_casewhen}
\alias{gen_casewhen}
\title{Generate the case-when list to use in the dplyr query}
\usage{
gen_casewhen(code_map, colname)
}
\arguments{
\item{code_map}{The result of calling gen_code_map()}

\item{colname}{The name of the column to use in the case-when statement}
}
\value{
A list of case when statements (expand with !!! in case_when)
}
\description{
Use this function to generate a list of case-when like statements ready
for use in a dplyr query to filter a diagnosis column by codes
}
