% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_map.R
\name{gen_cds_map}
\alias{gen_cds_map}
\title{Parse CDS technical output specification}
\usage{
gen_cds_map(tos, sheet, output = filename)
}
\arguments{
\item{tos}{The path to the technical output specification file}

\item{sheet}{The sheet name (corresponding to a database) to use}

\item{output}{The path to an output file for the generated mapping.yaml
This parameter is optional. If it is not specified, it will be
generated automatically based on the sheet name.}
}
\description{
This function parses the technical output specification for a
CDS dataset into a mapping.yaml file suitable for use
in mapped_server.
}
\details{
The technical output specification (TOS) is a document that
lists information about databases, such as column names and
documentation. Pass this file (and excel document) to this
function, along with the name of the sheet containing the
specification you want to parse.

In the generated output file, you will need to specify the
source database and tables names in order to map the file
to a particular database.
}
