% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{gen_clean_apc}
\alias{gen_clean_apc}
\title{Generate a synthetic admitted patient care database}
\usage{
gen_clean_apc(filename, seed = 1, nspells = 10)
}
\arguments{
\item{filename}{The name of the sqlite database file in which
to store the resulting database. The file will be placed in a
folder gendata/ in the working directory, which will be
created if it does not exist. Any files in gendata/ with the
same name will be overwritten. The filename can be anything,
but you may wish to give it an extension .db or .sqlite for
clarity}

\item{seed}{The seed to use for data generation. If all other
parameters are equal (excluding filename), then using the
same seed is guaranteed to produce the same dataset.}

\item{nspells}{The number of spells to include in the dataset}
}
\description{
Generate a synthetic table of admitted patient care
data and save it to an SQLite database file.
}
