% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_map.R
\name{gen_swd_map}
\alias{gen_swd_map}
\title{Read SWD guide}
\usage{
gen_swd_map(spec, output = filename)
}
\arguments{
\item{spec}{The SWD guide document (.xlsx)}

\item{output}{The file path to write the output}

\item{sheet}{The name of the sheet to read}
}
\description{
Read the BNSSG system-wide dataset guide document into a mapping file.
}
\details{
This function expects to find a sheet called "Data Tables and Fields",
containing a table of column names in the tables with documentation.
The table should have four columns: \verb{Table Name}, \verb{Column Name},
\verb{Data Type} and \code{Description}.
}
