% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata list}
\usage{
get_metadata(hash)
}
\arguments{
\item{hash}{The hash of the cache entry}
}
\description{
Get the metadata associated with a hash as a list
}
