% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{get_parse_stats}
\alias{get_parse_stats}
\title{Get ICD-10 parse statistics}
\usage{
get_parse_stats(x)
}
\arguments{
\item{x}{The icd10 codes list to process}
}
\value{
A list contains counts of each class
}
\description{
Get the parse statistics for a vector of icd10
codes. The statistics include the number of valid
and invalid codes, codes with trailing matter, and
empty strings.
}
