% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{get_types}
\alias{get_types}
\title{Get a character vector or types of icd10 codes}
\usage{
get_types(x)
}
\arguments{
\item{x}{The input vector of icd10 codes to parse}
}
\value{
A character vector or types
}
\description{
Returns a character vector containing single letters that
represent the type (the parse status) of the icd10 code.
The valid types are: C = code correctly parsed; X = code
is invalid; E = code is empty string; T = code parsed
correctly, but contains trailing matter.
}
