% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{\%in_group\%}
\alias{\%in_group\%}
\title{Binary operator version of in_groups}
\usage{
x \%in_group\% group
}
\arguments{
\item{x}{The icd10 codes (icd10 vector) to test}

\item{group}{A character vector of groups. If there
are multiple groups, codes will be returned for all
the groups listed.}
}
\description{
Binary operator version of in_groups
}
