% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{grep,Server-method}
\alias{grep,Server-method}
\title{Search the tables and columns in a database for partial names}
\usage{
\S4method{grep}{Server}(srv, col_pattern, tab_pattern)
}
\arguments{
\item{col_pattern}{The pattern to match column names (regexp)}

\item{tab_pattern}{The pattern to match table names (regexp)}

\item{db}{The Databases object to query}
}
\description{
Search the tables and columns in a database for partial names
}
