% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{group_string}
\alias{group_string}
\title{Extract groups character vector from icd10 object}
\usage{
group_string(x)
}
\arguments{
\item{x}{The the vector of icd10 objects to obtain groups
from}
}
\description{
Extract groups character vector from icd10 object
}
