% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{has_tag}
\alias{has_tag}
\title{Filter by tag}
\usage{
has_tag(value, tag)
}
\arguments{
\item{value}{A code string of the form foo.bar\<tag\>.some}

\item{tag}{A tag string, defined by a codes definition file}
}
\value{
TRUE if tag is present, false otherwise
}
\description{
Function to filter by the presence of a tag in a clinical code
column. Suitable for use as the data masking argument in dplyr
filter.
}
\details{
Clinical codes have the form foo.bar\<tag\>.some, where the token
inside the \<\> is called a tag string. This function returns TRUE if
\<tag\> is present in the code, otherwise it returns FALSE.
}
