% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10}
\alias{icd10}
\title{Create a new icd10 object}
\usage{
icd10(
  str = character(),
  codes_file = system.file("extdata", "code_groups/icd10.yaml", package = "icdb")
)
}
\arguments{
\item{str}{The input character vector of ICD code strings to parse}

\item{codes_file}{The reference codes file (defining codes and groupings)}
}
\value{
A new icd10 vector
}
\description{
An icd10 object is a vector class containing parsed ICD codes,
along with optional groups defined by a codes mapping file. The
codes are parsed from an input string, which contains ICD codes
that may have leading/trailing whitespace, may not contain
complete ICD codes, or may contain invalid codes. These codes
are categorised as part of the parsing process.
}
