% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10_gen_indices}
\alias{icd10_gen_indices}
\title{Index a codes definition structure}
\usage{
icd10_gen_indices(codes)
}
\arguments{
\item{codes}{The input nested list of codes}
}
\value{
The codes structure with indices (a nested list)
}
\description{
To facilite searching for codes in the configuration
file, it is important for each object (category or code)
to store a range of codes that it contains. R supports
lexicographical comparison of characters by default,
so all that is required is to store a pair representing
the start of the range and the end of the range. This
function adds an index key to the structure passed as
the argument.
}
\details{
The index key represents the first allowable code in
the category. This function replaces the code key in
the leaf nodes with an index, which contains the code
at that level.
}
