% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10_indices_to_code}
\alias{icd10_indices_to_code}
\title{Convert indices lists to ICD-10 codes}
\usage{
icd10_indices_to_code(indices, codes)
}
\arguments{
\item{indices}{A list of indices (itself a list)}

\item{codes}{The codes definition structure}
}
\value{
The named list containing the corresponding code
or category for these indices
}
\description{
Convert a list of indices to a list of codes
}
