% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10_sort_by_index}
\alias{icd10_sort_by_index}
\title{Reorder the child categories in a category}
\usage{
icd10_sort_by_index(cat)
}
\arguments{
\item{cat}{The category to reorder}
}
\value{
The modified category
}
\description{
The codes structure must be ordered by index at
every level. Most levels is already ordered
by category, which is fine except for the
chapter level (where the numerical order
of Roman numerals does not coincide with
the lexicographical order). Another exception
is U occuring after Z. The function takes
a category and sorts the child array (if it
exists), retuning the modified category.
}
\details{
This function calls itself recursively and
modifies all child categories inside the
argument cat.
}
