% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10_str_to_indices}
\alias{icd10_str_to_indices}
\title{Identify an ICD-10 code.}
\usage{
icd10_str_to_indices(str, codes)
}
\arguments{
\item{str}{The code string (from the database) to
search for (with dots removed)}

\item{codes}{The codes definition structure}
}
\value{
A vector of indices identifying the code
}
\description{
Search the ICD-10 codes definition structure
to find the indices that identify a particular
ICD-10 code.
}
\details{
An ICD10 string is a four-character
code like "C71.0". In the database, the
dot may be missing, there may be a trailing
dash (for filler codes, which should be X,
or any other high-level category code),
and there may be trailing matter (such as
dagger or asterisk codes).

Even though all the information about a code
is present in the string, it is necessary to
search the codes file to find the location
of the code.

Each level of the codes file has a category
or a codes key, with structure as follows:

category -> category +  -> category -> code
(chapter)   (code_range)   (triple)    (ICD-10)
I           A00-A09        A00         A00.0

The + after the code range indicates that
there may be an arbitrary number of code range
levels, involving increasingly nested levels
of the codes file. A code_range is identified
by the presence of a dash in the category.
}
