% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd_api_fetch_all}
\alias{icd_api_fetch_all}
\title{Fetch ICD-10 codes}
\usage{
icd_api_fetch_all(token, release = "2016", dir = ".")
}
\arguments{
\item{token}{The access token for the API}

\item{release}{The release year}

\item{dir}{Where to put the output files}
}
\description{
Fetch all the ICD10 code chapters into separate files.
}
\details{
The function outputs files of the form icd10_I.yaml
where I is replaced with the chapter number.
}
