% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd_api_get_codes}
\alias{icd_api_get_codes}
\title{Generate a codes file from the ICD-10 API}
\usage{
icd_api_get_codes(
  token,
  release = "2016",
  item = "I",
  endpoint = paste(root, release, item, sep = "/")
)
}
\arguments{
\item{token}{The access token obtained from icd_api_token}

\item{release}{The release year (default "2016"; Options "2019",
"2016", "2010", "2008")}

\item{item}{The item to retrieve; a string appended to the end of the
endpoint URL. This can be a chapter or a lower level item.}

\item{endpoint}{}
}
\value{
A leaf or category object (named list)
}
\description{
Use this function to automatically process the ICD-10 classifications into
a codes definition file that can be used to process diagnosis code columns.
This function uses the 2016 release of ICD-10.
}
\details{
Use this function to get each chapter of the ICD codes, as a nested list
object. For now, getting each chapter separately is a better way to avoid
the token timing out. The results can be combined to form a full ICD-10
file.
}
