% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd_api_request}
\alias{icd_api_request}
\title{Make a request to the WHO ICD API}
\usage{
icd_api_request(token, endpoint, data = list())
}
\arguments{
\item{token}{The authentication token (containing access_token key)}

\item{endpoint}{The URL endpoint}

\item{data}{A named list of the headers to send}
}
\value{
A named list of the contents of the request
}
\description{
After authenticating, make a request to an API endpoint. Use
}
\details{
Note: this function can be mocked for testing.
}
