% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{icdb_example}
\alias{icdb_example}
\title{Generate example data}
\usage{
icdb_example()
}
\description{
Generating example data from use in the documentation and
for experimenting with how the library works. Data is put
in a folder called gendata in the current working directory.
}
\details{
Any documentation examples that require data should call this
function before using the data in the gendata directory.
}
