% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{in_group}
\alias{in_group}
\title{Check if ICD-10 code is in a group}
\usage{
in_group(x, group)
}
\arguments{
\item{x}{The icd10 code (icd10 S3 class) to test}

\item{group}{The group as a string}
}
\value{
TRUE if the code is in the group, false otherwise
}
\description{
Check whether an icd10 code is in a particular group.
Suitable for use in the data masking arguments of
dplyr calls
}
