% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{is_valid}
\alias{is_valid}
\title{Return which ICD codes are valid as a logical vector}
\usage{
is_valid(x)
}
\arguments{
\item{x}{The icd10 codes to read}
}
\value{
A logical vector containing which codes are valid
}
\description{
Return which ICD codes are valid as a logical vector
}
