% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_empty_metadata}
\alias{make_empty_metadata}
\title{Make an empty level 1 metadata tibble}
\usage{
make_empty_metadata()
}
\value{
The empty cache tibble
}
\description{
This function creates an empty level 1 cache tibble, containing
the the columns: hash (for the unique key for the item);
data (the thing that is hashed, e.g. an sql query string);
hits (the number of times the entry was used); write_time;
last_access; and time (how long the original computation of
the cached data took).
}
