% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{make_table_getter}
\alias{make_table_getter}
\title{Make a table getter}
\usage{
make_table_getter(con, id)
}
\arguments{
\item{con}{The connection to the database}

\item{id}{The id (from DBI::dbListObjects) referencing the table}
}
\value{
A function which, when called, returns a dplyr::tbl
}
\description{
Make a function that returns a table getter. The function which
is returned can be called to produce a dplyr::tbl.
}
\details{
Even though the code uses () to access the table name, this level
of indirection is still necessary because of the bug referenced
in the build_object_tree body -- using a table getter means that
the tbl is only created when the user asks for it, removing some
edge cases with "bad" tables (like COLUMNS_EXTENSIONS in mysql
information_schema, which contains JSON columns).
}
