% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MappedSrv}
\alias{MappedSrv}
\title{Create a mapped databases (logical database object)}
\usage{
MappedSrv(
  ...,
  mapping = system.file("extdata", "mapping.yaml", package = "icdb")
)
}
\arguments{
\item{...}{Any arguments required by Server (see ?Server)}

\item{mapping}{the path to a mapping.yaml file}
}
\value{
A new MappedDB object
}
\description{
Create a new mapped database object. A mapped database is an object that contains
logical databases, logical tables and logical column names, with documentation.
The structure of the database is defined by a mapping.yaml file, which describes
how to obtain the fields from an underlying data source.
}
\details{
To make a mapped database, pass the Data Source Name (see Databases documentation)
along with a mapping.yaml file.
}
