% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MappedTab}
\alias{MappedTab}
\title{Create a MappedTab object}
\usage{
MappedTab(tab, logical_columns)
}
\arguments{
\item{tab}{The underlying Tab to use}

\item{logical_columns}{The columns portion of the parsed yaml config file}
}
\value{
A new MappedTab object
}
\description{
Create a mapped table object. This object stores the table itself
(a Tab), along with metadata providing the logical column names,
which underlying physical columns make up these logical columns,
documentation for the columns, and reduce strategies for collapsing
each logical column into one column.
}
