% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{new_icd10}
\alias{new_icd10}
\title{Make an ICD-10 object from a string}
\usage{
new_icd10(
  str = character(),
  codes = system.file("extdata", "icd10/icd10.yaml", package = "icdb")
)
}
\arguments{
\item{str}{The input string to parse}

\item{codes}{The ICD-10 codes definition file to use}
}
\value{
The new icd10 S3 object
}
\description{
Create a new icd10 (S3) object from a string
}
