% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_icd10_impl}
\alias{new_icd10_impl}
\title{Implementation of the ICD-10 parser for vectors of ICD-10 strings}
\usage{
new_icd10_impl(str, code_def)
}
\arguments{
\item{str}{The input character vector of strings that should be
parsed. The strings can have leading and trailing whitespace,
which will be removed before parsing by this function.}

\item{code_def}{The code definition structure, which is a nested
list of lists following the structure of the icd10.yaml file.}
}
\value{
A named list containing indices, type and groups
}
\description{
This function is the faster implementation of the ICD-10 parser.
The input is a vector of strings to be parsed, and the output is
a list of three vectors (the same length as the input) containing
information about the parsed codes.
}
