% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{new_Node}
\alias{new_Node}
\title{Node object for intermediate positions in the object tree.}
\usage{
new_Node(subobjects, ..., class = character())
}
\arguments{
\item{subobjects}{The list of Nodes or Tables in this Node}

\item{...}{Other parameters for derived classes}

\item{class}{The class parameter for derived class}
}
\value{
A new Node S3 object
}
\description{
The Server object is a nested list of Nodes, which each
represent an object in the database. This is the S3 class
for a Node.
}
\details{
The one job of the Node is to correctly handle the \code{$}
overload for the object being accessed in the object
tree. For the Server, \code{$} is overloaded to list the top
level objects. All other objects in the tree are either
Nodes or Tables. The \code{$} operator for a Node is overloaded
in such a way that it checks whether the node below it is
a Node or a Table. If it is a Node, it prints a list of the
contents of that Node. If it is a Table, it automatically
fetches the dplyr::tbl.

The Node object is a named list of either Nodes or Tables.
}
