% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{new_Table}
\alias{new_Table}
\title{Make a Table}
\usage{
new_Table(tbl, ..., class = character())
}
\arguments{
\item{tbl}{The underlying tbl to use}
}
\value{
The new Table object
}
\description{
Make a new Table object, which is the basic type used to
store tables in the library. This function should be
used in the table_getter to give a tibble object back
to the user.
}
