% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nhs_number_valid}
\alias{nhs_number_valid}
\title{Check NHS number validity}
\usage{
nhs_number_valid(nhs_number)
}
\arguments{
\item{nhs_number}{The NHS number (character) to check}
}
\value{
TRUE for a valid number, false otherwise
}
\description{
Check whether an NHS number, expressed as a character,
is valid.
}
