% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{parse_codes}
\alias{parse_codes}
\title{Get a flat list of codes}
\usage{
parse_codes(codes)
}
\arguments{
\item{codes}{The input codes definition list of lists}
}
\value{
A named character vector containing the codes
}
\description{
Get all the codes in a codes definition yaml file as a list.
}
\details{
A codes definition file is a yaml with a tree-like structure
where leaves of the tree are ICD codes, and non-leaf nodes
are categories. Each level in the structure contains either
a "categories" field (for a non-leaf) or a "code" field
(for a leaf). All levels in the structure contains a "docs"
field describing what the level means.
}
