% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{parse_icd10}
\alias{parse_icd10}
\title{Parse ICD-10 codes}
\usage{
parse_icd10(path, output = "icd10.yaml")
}
\arguments{
\item{path}{The input file path (tab separated)}

\item{output}{The filename of the output definition file}
}
\description{
Parse an ICD-10 codes file from the NHS Digital (TRUD) from the
"NHS ICD-10 5th Edition data files" item. The function converts the
codes to a codes definition file suitable for use in a mapped
server.
}
