% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{prune_level1}
\alias{prune_level1}
\title{Prune the level 1 cache}
\usage{
prune_level1()
}
\description{
If the level 1 cache is full, move older entries to the level 2 cache.
The element that is moved is the oldest element.
}
