% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{queries}
\alias{queries}
\title{Global variable storing the query cache folder path}
\format{
An object of class \code{Cache} of length 1.
}
\usage{
queries
}
\description{
Note: This was previously called cache, but that conflicted with the Cache
object (can't have upper and lower case filenames in package). Need to think
of a better name, but this will do for now. Might get rid of the object
entirely.
}
\keyword{datasets}
