% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{query_cache_path}
\alias{query_cache_path}
\title{Global variable storing the query cache folder path}
\format{
An object of class \code{character} of length 1.
}
\usage{
query_cache_path
}
\description{
Global variable storing the query cache folder path
}
\keyword{datasets}
