% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_nhs_number}
\alias{random_nhs_number}
\title{Make a random NHS number}
\usage{
random_nhs_number()
}
\value{
The NHS number as a character vector
}
\description{
This function makes a random, invalid, 10-digit NHS number
and returns it as a character.
}
\details{
An NHS number is a 10-digit numeric identifier that
satisfies a checksum condition.
}
