% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{raw,Database,character-method}
\alias{raw,Database,character-method}
\title{Perform an SQL query by directly passing the SQL string}
\usage{
\S4method{raw}{Database,character}(db, query)
}
\arguments{
\item{db}{The Database to submit to query to}

\item{query}{The query to submit (character string)}
}
\description{
Perform an SQL query by directly passing the SQL string
}
