% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{read_include}
\alias{read_include}
\title{Read an included config file.}
\usage{
read_include(include_file)
}
\arguments{
\item{include_file}{The path to the config file}
}
\value{
The nested list for the data stored in the included file
}
\description{
Read an included configuration file. The search path is the
current working directory first, followed by the extdata
folder.
}
\details{
This is used for the include key in the config files. This function
represents a generic approach to opening files, which will
search the working directory first and then extdata. All files
will be found this way. May want to consider modifying this
mechanism later, in which case it will be good to isolate it in
a few functions like this (the other one is currently get_codes,
which doesn't quite work like this and needs changing).
}
