% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{readCache}
\alias{readCache}
\title{Read an object from the cache}
\usage{
readCache(cache, data)
}
\arguments{
\item{cache}{The Cache object to read from}

\item{data}{The same data object passed to the writeCache function}
}
\value{
Returns the object associated with the data. Null if not found.
}
\description{
Use this function to get an object from
}
