% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{reduce,MappedTab-method}
\alias{reduce,MappedTab-method}
\title{Reduce a MappedTab}
\usage{
\S4method{reduce}{MappedTab}(x)
}
\arguments{
\item{x}{The MappedTab object to reduce}
}
\value{
A dplyr::tbl with the logical columns in the MappedTab
}
\description{
Reduce a mapped table object to a tbl with single logical columns
}
\details{
When a MappedTab is generated, each logical column corresponds to
a set of columns in the tbl. Before the table can be used, these
columns need to be reduced to a single logical column. See the
strategy.R file for more information.
}
