% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{searchCols,Database-method}
\alias{searchCols,Database-method}
\title{Search the tables and columns in a database for partial names}
\usage{
\S4method{searchCols}{Database}(db, col_pattern, tab_pattern)
}
\arguments{
\item{db}{The Database object to query}

\item{col_pattern}{The pattern to match column names (regexp)}

\item{tab_pattern}{The pattern to match table names (regexp)}
}
\description{
Search the tables and columns in a database for partial names
}
