% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{searchCols}
\alias{searchCols}
\title{Search the tables and columns in a database for partial names}
\usage{
searchCols(db, col_pattern, tab_pattern)
}
\arguments{
\item{db}{The Database object to query}

\item{col_pattern}{The pattern to match column names (regexp)}
}
\description{
Several of the databases are very large, with hundreds of tables each
with hundreds of columns. This function is designed to help with finding
relevant columns in the tables of the database, by partially matching the
table and column names and printing the results.
}
