% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\docType{class}
\name{Server-class}
\alias{Server-class}
\title{Server class wrapping an SQL server connection}
\description{
Server class wrapping an SQL server connection
}
\section{Slots}{

\describe{
\item{\code{con}}{Microsoft SQL Server.}

\item{\code{.S3Class}}{The inherited Node object}
}}

