% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{Server}
\alias{Server}
\title{Construct a Server object}
\usage{
Server(data_source_name = NULL, config = NULL, interactive = TRUE)
}
\arguments{
\item{data_source_name}{The data source name (DSN) for the database.  If
this argument is provided, it will be preferred to the config file.}

\item{config}{The path of a configuration file (YAML format)
containing database connection information and credentials. The default
value is the credential file stored in the inst/ directory.}

\item{interactive}{This parameter specifies whether the list of databases
and tables is populated. This is used to make the MappedSrv object load
faster, by only fetching the tables specified in the mapping.}
}
\value{
A new (S4) Server object
}
\description{
This object manages a connection to a database server. The connection
is configured using either a data source name, or a YAML file which stores
configuration information and credentials. After this function has run
without errors, you should have a new Server object containing a valid
connection. The data source is specified by a data source name, which is
configured using a Windows program. This is the preferred connection method.
}
\details{
A data source name (DNS) is a string that refers to a connection set
using the "ODBC data sources" app on Windows (8). See the package
documentation for more information about how to set this up.

If you cannot make the DSN work, you can also supply a configuration file,
in YAML format. Three keys are required: "driver" must be set to the
database driver (potentially "SQL Server"); "server" must be set the to
server name; and "database" must be set to the name of the database to
connect to.
}
