% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{methods}
\name{show,Database-method}
\alias{show,Database-method}
\alias{show-Database}
\alias{show,}
\alias{Database-method}
\title{Print out the Database object}
\usage{
\S4method{show}{Database}(object)
}
\arguments{
\item{object}{The object to be printed}
}
\description{
Print out the Database object
}
