% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{methods}
\name{show,Databases-method}
\alias{show,Databases-method}
\alias{show-Databases}
\alias{show,}
\alias{Databases-method}
\title{Print out the Databases object}
\usage{
\S4method{show}{Databases}(object)
}
\arguments{
\item{object}{The object to be printed}
}
\description{
Print out the Databases object
}
