% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{show,DatabaseS4-method}
\alias{show,DatabaseS4-method}
\title{Print out the Database object}
\usage{
\S4method{show}{DatabaseS4}(object)
}
\arguments{
\item{object}{The object to be printed}
}
\description{
Print out the Database object
}
