% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{show,mapped_server-method}
\alias{show,mapped_server-method}
\title{Print the mapped_server object}
\usage{
\S4method{show}{mapped_server}(object)
}
\arguments{
\item{object}{mapped_server to print}
}
\description{
Print the mapped_server object
}
