% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{show,MappedSrv-method}
\alias{show,MappedSrv-method}
\title{Print the MappedSrv object}
\usage{
\S4method{show}{MappedSrv}(object)
}
\arguments{
\item{object}{MappedSrv to print}
}
\description{
Print the MappedSrv object
}
