% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{show,MappedTab-method}
\alias{show,MappedTab-method}
\title{Show a mapped table object.}
\usage{
\S4method{show}{MappedTab}(object)
}
\arguments{
\item{object}{The object to show}
}
\description{
Print the contents of a mapped table object
}
\details{
This shows the associated dplyr::tbl and also the logical
column names with documentation
}
