% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\docType{methods}
\name{show,server-method}
\alias{show,server-method}
\alias{show-Databases}
\alias{show,}
\alias{Databases-method}
\alias{TODO:}
\alias{This}
\alias{function}
\alias{needs}
\alias{improvement}
\alias{--}
\alias{at}
\alias{the}
\alias{moment}
\alias{it}
\alias{is}
\alias{really}
\alias{just}
\alias{a}
\alias{placeholder}
\alias{for}
\alias{something}
\alias{more}
\alias{useful.}
\title{Print out the Databases object}
\usage{
\S4method{show}{server}(object)
}
\arguments{
\item{object}{The object to be printed}
}
\description{
Print out the Databases object
}
