% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{show,TableGetter-method}
\alias{show,TableGetter-method}
\title{Print the TableGetter}
\usage{
\S4method{show}{TableGetter}(object)
}
\arguments{
\item{object}{The TableGetter object}
}
\description{
Print a warning if the user tries to view the TableGetter object directly
}
