% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{sql_query}
\alias{sql_query}
\title{Submit an SQL query to a database object}
\usage{
sql_query(db, query)
}
\arguments{
\item{db}{The database to query}

\item{query}{A string with the SQL query}
}
\value{
A tibble containing the results
}
\description{
Submit an SQL query to a database object
}
