% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{sqlFromFile,server,character-method}
\alias{sqlFromFile,server,character-method}
\title{Perform a SQL query from a file}
\usage{
\S4method{sqlFromFile}{server,character}(db, file)
}
\arguments{
\item{db}{The Databases to submit to query to}

\item{file}{The file containing the query to submit}
}
\description{
If you have a file of SQL, submit it as a query to the database using this
function. It reads the file into a string and passes it to the sql_query
function. This function also caches the results of the query in a file
so that next time the function is called, the query will attempt to use
the cached file before using the SQL server. See the documentation for
sql_query for more information about how this works.
}
