% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlFromFile}
\alias{sqlFromFile}
\title{Submit an SQL query from a file and get the results}
\usage{
sqlFromFile(db, file)
}
\arguments{
\item{db}{The Databases object to submit the query to}

\item{file}{The path to the file containing SQL}
}
\value{
A tibble containing the results
}
\description{
Submit an SQL query from a file and get the results
}
