% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlQuery,Database,character-method}
\alias{sqlQuery,Database,character-method}
\title{Perform an SQL query by directly passing the SQL string}
\usage{
\S4method{sqlQuery}{Database,character}(db, query)
}
\arguments{
\item{db}{The Database to submit to query to}

\item{query}{The query to submit (character string)}
}
\description{
Submit an SQL query to the Database and obtain the results of the query in
a tibble dataframe. The query results are cached in a file on the disk,
so that if do the same query a second time, the local results are used.
}
