% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlQuery}
\alias{sqlQuery}
\title{Submit an SQL query to a database object}
\usage{
sqlQuery(db, query)
}
\arguments{
\item{db}{The database to query}

\item{query}{A string with the SQL query}
}
\value{
A tibble containing the results
}
\description{
Submit an SQL query to a database object
}
