% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategy_coalesce}
\alias{strategy_coalesce}
\title{Reduce by coalescing columns}
\usage{
strategy_coalesce(tbl, name)
}
\arguments{
\item{tbl}{An input dplyr::tbl}

\item{name}{The logical column name}
}
\value{
The tbl after reducing the columns
}
\description{
Reduce a logical column by the coalesce strategy, which treats elements
of the logical column as expressing the same information. Physical columns
are ordered according to the ordering specified in the yaml file. The value
in the final logical column is selected from the first physical column for
which the value is not NULL. This is the same as the SQL COALESCE function,
which does the same thing.
}
\details{
TODO: consider replacing this function by the built-in coalesce function
which probably works. Secondly, it may be a good idea to put the default
statement of the case when to catch issues (like everything being null).
}
