% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategy_coalesce_exclude_null}
\alias{strategy_coalesce_exclude_null}
\title{Reduce by coalescing columns, excluding resulting null values}
\usage{
strategy_coalesce_exclude_null(tbl, name)
}
\arguments{
\item{tbl}{The tbl to process}

\item{name}{The logical column name}
}
\value{
The tbl after reducing
}
\description{
This is the same as the coalesce strategy, but it also filters out rows
where the coalesced result is null. This is important for logical
columns where a null result would render the entire row meaningless.
}
\details{
This function excludes all undefined value
}
