% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Tab-class}
\alias{Tab-class}
\title{Class to store just one table. Each of these objects goes
into a Tables list}
\description{
Class to store just one table. Each of these objects goes
into a Tables list
}
