% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{Tab}
\alias{Tab}
\title{Make a new Tab object}
\usage{
Tab(table_getter, docs_getter = NULL)
}
\value{
The new Tab object
}
\description{
Create a new Tab object
}
\details{
The purpose of the Tab object is to store information about a table
in Tables (Tab is used instead of Table to avoid confusion). The
object stores two things: a function to get the table, and a function
to get the documentation for the table.
}
