% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Table-class}
\alias{Table-class}
\title{Table class}
\description{
Class for wrapping a database table
}
\details{
The only purpose of this at the moment is to allow overloading
the '$' operator for table access. There might be more uses for
it later.
}
