% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table,Database,character-method}
\alias{table,Database,character-method}
\title{Get the dplyr table wrapper for a table passed as a string}
\usage{
\S4method{table}{Database,character}(db, tab)
}
\arguments{
\item{db}{.}

\item{tab}{.}
}
\description{
Get the dplyr table wrapper for a table passed as a string
}
