% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table}
\alias{table}
\title{Get a table in the database in a form ready for dplyr processing}
\usage{
table(db, tab)
}
\arguments{
\item{tab}{The name of the Table object to use read}
}
\value{
A dply::tbl object which wraps a database table
}
\description{
Use this function to obtain a dplyr::tbl (a kind of shell object) from a
table in the database. This object does not contain the full data to begin
with. Instead, it can be used as the basis of dplyr-based SQL queries,
finishing with a \%>\% collect() call which does actually produce data. This
function is an alternative to using a direct sql query using the sql()
generic
}
