% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table_better_getter}
\alias{table_better_getter}
\title{The name isn't staying. This function is used for all the "good" databases
that actually list objects properly (everything except SQL Server). The
id field can be passed to dbQuoteIdentifier to create a query}
\usage{
table_better_getter(con, id)
}
\description{
The name isn't staying. This function is used for all the "good" databases
that actually list objects properly (everything except SQL Server). The
id field can be passed to dbQuoteIdentifier to create a query
}
