% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table_getter}
\alias{table_getter}
\title{Get a function which returns a table object}
\usage{
table_getter(srv, database, table_schema, table_name)
}
\arguments{
\item{srv}{The Server object to use (containing the connection)}

\item{database}{The database name}

\item{table_schema}{The table schema name}

\item{table_name}{The table name}
}
\description{
Function factory returning a function that gets a dplyr::tbl
}
\details{
This function is the way to associate a function with every table
object in the db list, and replaces the need to overload \code{$}, while
also avoiding the problem of storing the table name and database
name in the functions environment. This is the purpose of a function
factory; to capture variables in the enclosing environment and allow
them to persist when the function is called. Read this page and the
associated environment sections for a full explanation:
https://adv-r.hadley.nz/function-factories.html
}
