% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{table_node}
\alias{table_node}
\title{Make a table_node}
\usage{
table_node(tbl, ..., class = character())
}
\arguments{
\item{tbl}{The underlying tbl to use}

\item{...}{Further arguments from other constructors}

\item{class}{Other classes (used for subclasses)}
}
\value{
The new table_node object
}
\description{
Make a new table_node object, which is the basic type used to
store tables in the library. This function should be
used in the table_getter to give a tibble object back
to the user.
}
