% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\docType{class}
\name{TableGetter-class}
\alias{TableGetter-class}
\title{Simple wrapper to print a user error. This class inherits from
function, which means it is expecting to be called. The purpose
of wrapping this in a class is to throw an error in the show
method, to catch the case where a user tries to view a table
by just writing table name without (). Without this, the user
would just get the body of the function, which is not really very
helpful for debugging.}
\description{
Simple wrapper to print a user error. This class inherits from
function, which means it is expecting to be called. The purpose
of wrapping this in a class is to throw an error in the show
method, to catch the case where a user tries to view a table
by just writing table name without (). Without this, the user
would just get the body of the function, which is not really very
helpful for debugging.
}
