% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{to_icd10}
\alias{to_icd10}
\title{Convert character to icd10}
\usage{
to_icd10(vec)
}
\arguments{
\item{vec}{The input character vector to parse}
}
\value{
The icd10 vector (same length as input)
}
\description{
Convert a character vector to an icd10 vector
}
