% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{vec_ptype_abbr.icdb_icd10}
\alias{vec_ptype_abbr.icdb_icd10}
\title{For this to work, I needed to import the generic.
This feels unnecessary -- find the proper way to
set a method for a generic not in the class.}
\usage{
\method{vec_ptype_abbr}{icdb_icd10}(x, ...)
}
\arguments{
\item{x}{The object to abbreviate the name of}

\item{...}{Further parameters (todo: check what for)}
}
\description{
For this to work, I needed to import the generic.
This feels unnecessary -- find the proper way to
set a method for a generic not in the class.
}
