% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_cache}
\alias{write_cache}
\title{Write an element into the cache}
\usage{
write_cache(data, object, time)
}
\arguments{
\item{data}{Data about the object, used to generate the hash}

\item{object}{The main (typically large) object to store in the cache}

\item{time}{The duration of the operation that requires caching}
}
\description{
This function creates an entry in the cache. It should not be called by the
user of the package. The data comprises two parts: an object, which is
typically large, which is used to store "results" in the cache. The second
part is data about the results, which is used to generate a key. The
data can be anything, but a named list is a simple way to organise the
necessary data.
}
\details{
Data is stored in two places: the level 1 cache, which is stored in memory;
and the level 2 cache, which is written to disk. Identical data is stored in
both locations, in terms of access time, etc. This function initialises the
hits to 1, meaning that this is the first access. All subsequent accesses of
the data should use the readCache function, which modifies the metadata in the
cache.

This function will not check the cache before writing to it, so any duplicate
data will be overwritten. Only call this function after establish via
readCache that the data is not already present in the cache.
}
