% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_level2}
\alias{write_level2}
\title{Flush entry to the level2 cache}
\usage{
write_level2(metadata)
}
\arguments{
\item{metadata}{The metadata as a named list}
}
\description{
Write en entry from the leve1 1 cache to the level 2 cache
}
