% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{writeCache}
\alias{writeCache}
\title{Write an element into the cache}
\usage{
writeCache(cache, data, object)
}
\arguments{
\item{cache}{A Cache object in which to store the data}

\item{data}{Data about the object, used to generate the hash}

\item{object}{The main (typically large) object to store in the cache}
}
\description{
This function creates an entry in the cache. It should not be called by the
user of the package. The data comprises two parts: an object, which is
typically large, which is used to store "results" in the cache. The second
part is data about the results, which is used to generate a key. The
data can be anything, but a named list is a simple way to organise the
necessary data.
}
\details{
This function will not check the cache before writing to it, so any duplicate
data will be overwritten. Only call this function after establish via
readCache that the data is not already present in the cache.
}
