% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{build_object_tree}
\alias{build_object_tree}
\title{Build the tree of accessible database object}
\usage{
build_object_tree(con, prefix)
}
\arguments{
\item{con}{The database connection object}

\item{prefix}{The starting prefix defining the root of the tree (ID type)}
}
\value{
Nested list structure containing the object tree
}
\description{
Get the tree of accessible objects in the database connection
}
\details{
Objects in a database connection are stored as a tree. At the root of
the tree is a starting prefix, which may represent a collection of
databases, a database, or even a table in a database. This function
recursively traverses the objects in the database using DBI::dbListObjects()
and stores the results in a tree structure for the purpose of automcompleting
object names.

The entire object tree for a database server can be quite large. This
causes the database connection to lag when it is first opened. However, there
is no way to avoid this delay if the list-based-autocompleting method is
used, which is currently the only implemented option. Setting a specific
prefix may help alleviate this performance problem if the area of interest
is known.

It may be possible to set the default database prefix in the data source
setup (Windows).
}
