% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{Cache}
\alias{Cache}
\title{Create a new cache object}
\usage{
Cache(path = "cache/")
}
\arguments{
\item{path}{The directory in which to store the cache, defaults to "cache/" relative
to the current working directory.}
}
\value{
A new Cache object
}
\description{
Create a new cache object
}
