% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_r6.R
\name{Database}
\alias{Database}
\title{Construct a Database object}
\description{
Construct a Database object

Construct a Database object
}
\examples{

## ------------------------------------------------
## Method `Database$new`
## ------------------------------------------------


db <- Database$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Database-new}{\code{Database$new()}}
\item \href{#method-Database-clone}{\code{Database$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-new"></a>}}
\if{latex}{\out{\hypertarget{method-Database-new}{}}}
\subsection{Method \code{new()}}{
This object manages a database connection. The connection is configured
using a json file which stores configuration information and credentials.
The fields of this configuration file are as follows:
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$new(
  db_config = system.file("extdata", "db_config.json", package = "rdatabase")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_config}}{The absolute path of a configuration file (csv
format) containing database connection information and credentials.
The default value is the credential file stored in the inst/ directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
driver:
}

\subsection{Returns}{
A new (R6) Database object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
db <- Database$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Database-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
