% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Databases-class}
\alias{Databases-class}
\title{Overload $ for Tables so that it returns the dplyr::tbl}
\arguments{
\item{x}{The Tables object (element in a Databases object)}

\item{name}{The table name to get}
}
\value{
The dplyr::tbl for the selected table
Databases class wrapping an SQL server connection
}
\description{
Get the table with the specified name as a dplyr::tbl
}
\section{Slots}{

\describe{
\item{\code{connection}}{Microsoft SQL Server.}

\item{\code{config}}{list. Databases connection information as a named list}

\item{\code{dsn}}{Domain source name (Windows only)}

\item{\code{.Data}}{From the contained list}
}}

