% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get_codes}
\alias{get_codes}
\title{Parse multiple codes files}
\usage{
get_codes(codes_files)
}
\arguments{
\item{codes_files}{A list of filenames to parse}
}
\value{
A named list suitable for gen_code_map()
}
\description{
Parse multiple codes files and combine the result
}
