% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{get_tbl}
\alias{get_tbl}
\title{Get a dplyr::tbl corresponding to a table}
\usage{
get_tbl(srv, database, table)
}
\arguments{
\item{srv}{The Databases object containing the server connection}

\item{database}{The name of the database}

\item{table}{The name of the table}
}
\value{
The dplyr::tbl (shell) for the table
}
\description{
This is the function for obtaining a table in non-interactive Databases mode.
The purpose of this function is to speed up the initial loading of the
Databases object, in an non-interactive environment. However, that means that
it is necessary to recheck the database server for the schema name each
time a table is requested. This does not add too much overhead.
}
\details{
Currently, this function only supports Microsoft SQL server.
}
