% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd10_str_to_indices}
\alias{icd10_str_to_indices}
\title{Parse an ICD-10 code string}
\usage{
icd10_str_to_indices(str, codes, groups)
}
\arguments{
\item{str}{The string to parse}

\item{codes}{The reference codes structure}

\item{groups}{The current set of groups for this code}
}
\value{
A named list containing indices (a list)
and type (containing the status of the parse).
}
\description{
This function takes a string and searches the codes
structure to find a match, returning the list of
indices defining the location of the code within
the structure. The purpose of the indices is to
store them in an icd10 object, allowing information
about the code to be queried.
}
\details{
Return type codes: 0 for success; 1 for empty
string; 2 for invalid code;
}
