% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd_api_token}
\alias{icd_api_token}
\title{Get the API token}
\usage{
icd_api_token()
}
\value{
The token (named list)
}
\description{
Get the authentication token for the WHO ICD API
}
\details{
Call this function once to get the token, and then use it
in the icd10_api calls

Note: this function can be mocked for testing.
}
