% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{in_group}
\alias{in_group}
\title{Check if ICD-10 code is in a group}
\usage{
in_group(x, group)
}
\arguments{
\item{x}{The icd10 codes (icd10 vector) to test}

\item{group}{A character vector of groups. If there
are multiple groups, codes will be returned for all
the groups listed.}
}
\value{
TRUE where the code is in the group,
FALSE otherwise
}
\description{
Check whether an icd10 code is in a particular group.
Suitable for use in the data masking arguments of
dplyr calls. The function will throw an error if
it is supplied a column which is not class icd10.
}
