% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{new_node}
\alias{new_node}
\title{node object for intermediate positions in the object tree.}
\usage{
new_node(subobjects, ..., class = character())
}
\arguments{
\item{subobjects}{The list of nodes or table_nodes in this node}

\item{...}{Other parameters for derived classes}

\item{class}{The class parameter for derived class}
}
\value{
A new node S3 object
}
\description{
The server object is a nested list of nodes, which each
represent an object in the database. This is the S3 class
for a node.
}
\details{
The one job of the node is to correctly handle the \code{$}
overload for the object being accessed in the object
tree. For the server, \code{$} is overloaded to list the top
level objects. All other objects in the tree are either
nodes or table_nodes. The \code{$} operator for a node is overloaded
in such a way that it checks whether the node below it is
a node or a table_node. If it is a node, it prints a list of the
contents of that node. If it is a table_node, it automatically
fetches the dplyr::tbl.

The node object is a named list of either nodes or table_nodes.
}
