% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_cache}
\alias{read_cache}
\title{Read an object from the cache}
\usage{
read_cache(data)
}
\arguments{
\item{data}{The same data object passed to the writeCache function}
}
\value{
Returns the object associated with the data. Null if not found.
}
\description{
Use this function to get an object from
}
