% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\name{run}
\alias{run}
\title{Collect and cache SQL query results}
\usage{
run(x, lifetime = NULL, ...)
}
\arguments{
\item{x}{The dplyr SQL query to collect}

\item{lifetime}{The default amount of time that cache results
will remain valid. Specified as a lubridate duration (e.g.
lubridate::dhours(24)), with default value 24 hours.}

\item{...}{Other arguments for dplyr::collect()}
}
\value{
The query results
}
\description{
This function is a replacement for the dplyr::collect() function
that is used to submit an SQL query to a database. This function
is called in the same way, by piping to collect(), but it caches
the query behind the scenes, so that the next time the query is
fetched, the results are available quicker.
}
\details{
The cache is disabled by default. Make sure you call use_cache
(see the documentation) if you want to use the cache.

You can use the lifetime parameter to overwrite the cache lifetime
specified in the arguments to use_cache().
}
