% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srv.R
\docType{class}
\name{server-class}
\alias{server-class}
\title{server class wrapping an SQL server connection}
\description{
server class wrapping an SQL server connection
}
\section{Slots}{

\describe{
\item{\code{con}}{Microsoft SQL server.}

\item{\code{.S3Class}}{The inherited node object}
}}

