% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_s4.R
\name{show,DatabaseS4-method}
\alias{show,DatabaseS4-method}
\title{Print out the database connection object}
\usage{
\S4method{show}{DatabaseS4}(object)
}
\arguments{
\item{DatabaseS4}{}
}
\description{
Print out the database connection object
}
