% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategy_codes_from}
\alias{strategy_codes_from}
\title{Remap a column of codes to hierarchy of strings}
\usage{
strategy_codes_from(tbl, name, codes_file)
}
\arguments{
\item{tbl}{The tbl to process}

\item{name}{The logical column name}

\item{codes_file}{The name of the codes configuration file}
}
\value{
The tbl after reducing
}
\description{
Read ICD codes (or any other codes) from a codes configuration file and
remap the column values with readable hierarchical strings.
}
\details{
Note that this function only currently works on logical columns with
only one source column.
}
