% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Table-class}
\alias{Table-class}
\title{A Table class for a table in the Database class}
\description{
The purpose of this class is to wrap the dplyr table shell object for
storing in the Database class. The Database class is essentially a list
of these objects, named by the table name. This permits tab-completion of
table names when a database is created. However, the innards of this class
is only populated when the table is created, so that it is not necessary
to store thousands of dplyr::tab objects. There is more optimisation work
to do, but this will test the principle.
}
\section{Slots}{

\describe{
\item{\code{tabname}}{character The name of the table}
}}

