% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table_getter}
\alias{table_getter}
\title{Get a function which returns a table object}
\usage{
table_getter(db, database, tabname)
}
\arguments{
\item{db}{The database object to use (containing the connection)}

\item{database}{The database name}

\item{tabname}{The table name}
}
\description{
Function factory returning a function that gets a dplyr::tbl
}
\details{
This function is the way to associate a function with every table
object in the db list, and replaces the need to overload \code{$}, while
also avoiding the problem of storing the table name and database
name in the functions environment. This is the purpose of a function
factory; to capture variables in the enclosing environment and allow
them to persist when the function is called. Read this page and the
associated environment sections for a full explanation:
https://adv-r.hadley.nz/function-factories.html

This function needs a bit of work -- this is only a first draft.
The main issue is hard coding the dbo object, which should really
come from somewhere and is probably going to lead to bugs down the
line. When I find a failure case, it can be fixed here. Currently, the
function just forms the name database.dbo.tabname, which seems to work
fine. However, it would be better to understand what the catalog and
schema really are and make this robust.
}
