% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{use_cache}
\alias{use_cache}
\title{Turn the cache on or off}
\usage{
use_cache(state)
}
\arguments{
\item{state}{TRUE to turn the cache on, FALSE to turn it off}
}
\description{
Use this function to turn the cache on or off. If the cache is
on, then data will be written to and read from the cache. If
the cache is off, then the read_cache function will always
return NULL and the write cache function will return without
doing anything. Turning the cache off does not clear the
cache, and turning the cache back on later will make available
the same cache contents as before.
}
\details{
If caching is enabled, and you perform exactly the same query
again (for example, by running the same script), then the
cached results will be used, which will speed up the running
of the script. This can make development and debugging easier,
because the turnaround time for running commands and scripts
is reduced.

If caching is disabled, then results will always come from the
database. This can make code take longer to run, but will
always guarantee that results are up to date.

The cache is disabled by default. This is to ensure that, by
default, queries always return up-to-date results.
}
