% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_level2}
\alias{write_level2}
\title{Write entries to the level2 cache}
\usage{
write_level2(metadata)
}
\arguments{
\item{metadata}{The metadata as a tibble}
}
\description{
Write entries from the level 1 cache to the level 2 cache.
The entries are supplied a rows of a tibble.
}
